+++ exhibitera/changelog.md
@@ -16,6 +16,7 @@
 ### InfoStation
 - The button row is now scrollable for large numbers of buttons [Morgan Rehnberg]
 - Chane the size of buttons to fit your needs [Morgan Rehnberg]
+- Better handling of images when they are larger than the text. [David Morgan]
 
 ### Media Browser
 - New style options give you greater control over the look and feel of Media Browser [Morgan Rehnberg]